+++ src/semantic/checks/expression.rs
@@ -193,7 +193,54 @@ impl SemanticCheck for Expression<'_> {
                 // After all these checks, we are ok
                 Ok(())
             }
-            Expression::Array(_) => unimplemented!("check for array"),
+            Expression::Array(arr) => {
+                // There is some duplicated logic here with
+                // the typechecking of an array in the impl
+                // of the trait Typed
+
+                // The first expression found determines the
+                // type of the array
+
+                // TODO: for now an array must have at least 1 element,
+                // if the array is empty there is no way to know its type.
+                // TODO: if the array is empty then its
+                // datatype should be determined by its usage.
+                if arr.exps.is_empty() {
+                    return Err(MistiError::Semantic(SemanticError {
+                        error_start: arr.start,
+                        error_end: arr.end,
+                        reason: format!(
+                            "An array must have at least 1 element to determine its type. This will be fixed later."
+                        ),
+                    }));
+                }
+
+                let mut expressions = arr.exps.iter();
+                let first_expr = expressions.next().unwrap();
+                let first_type = first_expr.get_type(scope)?;
+
+                // then check that every expression has the same type
+                for exp in expressions {
+                    let exp_type = exp.get_type(scope)?;
+                    if !exp_type.equals(&first_type) {
+                        // TODO: subtyping
+
+                        // error, found an item with a diferent datatype
+                        let (error_start, error_end) = exp.get_position();
+                        return Err(MistiError::Semantic(SemanticError {
+                            error_start,
+                            error_end,
+                            reason: format!(
+                                "All elements of an array must have the same datatype. Expected {:?}, got {:?}",
+                                first_type,
+                                exp_type,
+                            ),
+                        }));
+                    }
+                }
+
+                Ok(())
+            },
         }
     }
 }