+++ src/semantic/checks/mod.rs
@@ -5,3 +5,4 @@ pub mod expression;
 pub mod for_loop;
 pub mod function_declaration;
 pub mod top_level_declaration;
+pub mod while_loop;